Fix namespaces and a small bug that appeared with refactor
Fix namespaces
