+++ pimi-connect-api/pimi-connect-api.API/Program.cs
@@ -1,10 +1,12 @@
 using FluentValidation;
 using FluentValidation.AspNetCore;
 using Microsoft.EntityFrameworkCore;
+using pimi_connect_api.Attachment;
 using pimi_connect_api.Hubs;
 using pimi_connect_api.Middleware;
 using pimi_connect_api.Services;
 using pimi_connect_api.Services.Interfaces;
+using pimi_connect_api.User;
 using pimi_connect_app.Data.AppDbContext;
 using pimi_connect_app.Data.Models;
 using pimi_connect_app.Data.Models.Validators;
