+++ pimi-connect-api/pimi-connect-api.API/User/UserController.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Mvc;
 using pimi_connect_api.Services.Interfaces;
 using pimi_connect_app.Data.Models;
 
-namespace pimi_connect_api.Controllers;
+namespace pimi_connect_api.User;
 
 [ApiController]
 [Route("api/[controller]")]
